# flask_app/swagger/openapi.yaml
swagger: 3.0
openapi: 3.0.0

info:
  description: Auth API
  version: "1.0.0"
  title: Auth API
tags:
  - name: 'auth_api'
    description: 'Authentication api'

paths:
  /api/v1/auth/signup:
    post:
      summary: Регистрация нового пользователя
      tags:
        - sign up
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        responses:
          '201':
            description: Пользователь создан
          '400':
            description: Регистрация не прошла

  /api/v1/auth/login:
    post:
      summary: 'Вход в учетную запись'
      tags:
        - login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
        responses:
          '200':
            description: 'Success login'
          '401':
            description: 'Auth failed'


  /api/v1/auth/logout:
    post:
      summary: Выход из учетной записи
      tags:
        - logout
      responses:
        '200':
          description: 'Success logout'
        '401':
          description: 'Auth failed'
      security:
        - AccessToken: [ ]
        - RefreshToken: [ ]

  /api/v1/auth/refresh:
    post:
      summary: 'Обновление токена'

  /api/v1/auth/change-password:
    post:
      summary: 'Изменить пароль'


components:
  securitySchemes:
      AccessToken:
        type: apiKey
        in: cookie
        name: API-KEY-ACCESS
      RefreshToken:
        type: apiKey
        in: cookie
        name: API-KEY-REFRESH

  schemas:
    User:
      type: object
      required:
        - login
        - password
        - first_name
        - last_name
      properties:
        login:
          type: string
        password:
          type: string
        first_name:
          type: string
        last_name:
          type: string
  